Extract the A4PageWrapper into a separate file.
